+++ node-express/apps/app-template/graphql/index.js
@@ -5,7 +5,7 @@ let graphql
 module.exports = function (app, server) {
   if (!graphql) {
     const schema = require('./schema')
-    graphql = require('@es-labs/node/express/graphql')(app, server, schema)
+    graphql = require('../../common/graphql')(app, server, schema)
   }
   return graphql
 }