+++ node-express/apps/app-template/tests/integration/category.route.test.js
@@ -11,7 +11,7 @@ let endpointUrl
 beforeAll(async () => {
   await require('@es-labs/node/config')(process.cwd())
   sqldb = await require('@es-labs/node/services/db/knex').open()
-  require('@es-labs/node/express/preRoute')(app, express, global.CONFIG)
+  require(APP_PATH + '/common/preRoute')(app, express, global.CONFIG)
   require(APP_PATH + '/router')(app)
 
   const { createToken } = require('@es-labs/node/auth')